[backend/frontend] Move module to internal
