+++ opencti-platform/opencti-graphql/src/domain/stixCoreObject.js
@@ -56,7 +56,7 @@ import { pushToConnector } from '../database/rabbitmq';
 import { now } from '../utils/format';
 import { ENTITY_TYPE_CONNECTOR } from '../schema/internalObject';
 import { deleteFile, storeFileConverter, upload } from '../database/file-storage';
-import { findById as documentFindById } from '../modules/document/document-domain';
+import { findById as documentFindById } from '../modules/internal/document/document-domain';
 import { elCount, elUpdateElement } from '../database/engine';
 import { generateStandardId, getInstanceIds } from '../schema/identifier';
 import { askEntityExport, askListExport, exportTransformFilters } from './stix';
